More unit tests
Move PriceCalculations to it's own file. Started Unit tests for class.
