+++ server/tests/Utility_Test.php
@@ -2,6 +2,15 @@
 declare(strict_types=1);
 //command> phpunit .\htdocs\Game-Library\server\tests
 
+/*
+Getting PHPunit to work in XAMPP:
+1) Download PHPunit.phar from this site
+   https://phpunit.de/getting-started/phpunit-9.html
+2) Copy the file to XAMPP/PHP folder
+3) Rename or delete the existing phpunit file (no extention)
+4) Rename the downloaded file to phpunit (no extention)
+*/
+
 use PHPUnit\Framework\TestCase;
 
 // We require the file we need to test.
